# src/main/java/fr/esgi/ecommerce/dto/review/ReviewResponse.java
package fr.esgi.ecommerce.dto.review;

import lombok.Data;

import java.time.LocalDate;

@Data
public class ReviewResponse {
    private Long id;
    private String author;
    private String message;
    private Integer rating;
    private LocalDate date;
}